+++ router/java/src/net/i2p/router/networkdb/kademlia/RepublishLeaseSetJob.java
@@ -44,11 +44,11 @@ public class RepublishLeaseSetJob extends JobImpl {
     private static final String PROP_TIMEOUT = "router.leaseSetPublishTimeout";
     private static final String PROP_RETRY_DELAY = "router.leaseSetPublishRetryDelay";
     private static final String PROP_MAX_RETRY_DELAY = "router.leaseSetPublishMaxRetryDelay";
-    public final static long REPUBLISH_LEASESET_TIMEOUT_DEFAULT = 60 * 1000;
-    public final static int RETRY_DELAY_DEFAULT = 20 * 1000;
-    public final static int RETRY_MAX_DELAY_DEFAULT = 30 * 1000;
-    private final static long EXPIRY_WINDOW = 3 * 60 * 1000;
-    private static final long CACHE_CLEANUP_THRESHOLD = 15 * 60 * 1000;
+    public final static long REPUBLISH_LEASESET_TIMEOUT_DEFAULT = 60L * 1000;
+    public final static int RETRY_DELAY_DEFAULT = (int) (20L * 1000);
+    public final static int RETRY_MAX_DELAY_DEFAULT = (int) (30L * 1000);
+    private final static long EXPIRY_WINDOW = 3L * 60 * 1000;
+    private static final long CACHE_CLEANUP_THRESHOLD = 15L * 60 * 1000;
     private static final ConcurrentHashMap<Hash, Boolean> _retryInProgress = new ConcurrentHashMap<>();
     private static final ConcurrentHashMap<Hash, Long> _lastPublishLogTime = new ConcurrentHashMap<>();
     private static final ConcurrentHashMap<Hash, Long> _lastVerifyLogTime = new ConcurrentHashMap<>();
@@ -105,8 +105,8 @@ public class RepublishLeaseSetJob extends JobImpl {
                 _facade.stopPublishing(_dest);
                 return;
             }
-            if (uptime < 5 * 1000) {
-                long delay = Math.max(1000, 5 * 1000 - uptime);
+            if (uptime < 5L * 1000) {
+                long delay = Math.max(1000, 5L * 1000 - uptime);
                 scheduleRepublish(delay);
                 return;
             }
@@ -137,7 +137,7 @@ public class RepublishLeaseSetJob extends JobImpl {
                             }
                             lastPubLog = null; // Force log
                         }
-                        if (_log.shouldInfo() && (lastPubLog == null || (now - lastPubLog > 10 * 1000))) {
+                        if (_log.shouldInfo() && (lastPubLog == null || (now - lastPubLog > 10L * 1000))) {
                             _log.info("Publishing LeaseSet" + name + " [" + _dest.toBase32().substring(0,8) +
                                        "] (expires in " + (timeUntilExpiry / 1000) + "s)...");
                             _lastPublishLogTime.put(_dest, now);
@@ -200,7 +200,7 @@ public class RepublishLeaseSetJob extends JobImpl {
     }
 
     private long getRepublishInterval() {
-        return getContext().getProperty("i2p.netdb.republishInterval", 5*60*1000);
+        return getContext().getProperty("i2p.netdb.republishInterval", 5L * 60 * 1000);
     }
 
     private void scheduleRepublish(long delayMs) {
@@ -336,7 +336,7 @@ public class RepublishLeaseSetJob extends JobImpl {
                 clearRetryInProgress();
                 long now = getContext().clock().now();
                 Long lastNotRequeueLog = _lastNotRequeueLogTime.get(_ls.getHash());
-                if (_log.shouldInfo() && (lastNotRequeueLog == null || (now - lastNotRequeueLog > 10 * 1000))) {
+                if (_log.shouldInfo() && (lastNotRequeueLog == null || (now - lastNotRequeueLog > 10L * 1000))) {
                     _log.info("Not requeueing LeaseSet" + name + " [" +
                               _ls.getDestination().calculateHash().toBase32().substring(0,8) +
                               "] -> Newer LeaseSet exists locally");
@@ -366,7 +366,7 @@ public class RepublishLeaseSetJob extends JobImpl {
                     requeueRepublish();
                     return;
                 }
-                if (_log.shouldInfo() && (lastVerifyLog == null || (now - lastVerifyLog > 10 * 1000))) {
+                if (_log.shouldInfo() && (lastVerifyLog == null || (now - lastVerifyLog > 10L * 1000))) {
                     _log.info("Verifying LeaseSet publication" + name + " [" +
                               _ls.getDestination().calculateHash().toBase32().substring(0,8) + "] via floodfill...");
                     _lastVerifyLogTime.put(_ls.getHash(), now);
@@ -413,7 +413,7 @@ public class RepublishLeaseSetJob extends JobImpl {
                 }
             };
 
-            _facade.lookupLeaseSetRemotely(_ls.getHash(), onFound, onFailed, 10*1000, null);
+            _facade.lookupLeaseSetRemotely(_ls.getHash(), onFound, onFailed, 10L * 1000, null);
         }
     }
